google/slippy: Enable LTE PCIe on NGFF socket and SIM USB
The Acer C720(P) has been designed to have a LTE modem on a NGFF connector.

Unfortunately the final revision of this chromebook doesn't have the connector
populated, for this reason on Coreboot this functionality is not present.

Adding back this feature is not easy, but not impossible.
This is a first path to:
 - Enable the USB port on the SIM tray. (USB #5)
 - Enable the PCIE_2 to provide a full working M.2 connector.

The NGFF connector have the following features:
 - USB 2.0
 - UIM
 - PCIE 1x

On the mainboard there are a few component that are not populated
some of those can be easily soldered (Like the NGFF connector) others
can be added as daughter board.

Most of the work must be done on the Power Gate. Like bypass capacitors.
Looks like that all the 10K pullups and 0Ohm links are present in the board.

The SIM tray socket is not populated, the part no. is unknown for that reason I have
made a custom board with all the required components.

This NGFF may be used to add more storage to the chromebook.

Signed-off-by: Nicolò Veronese <nicveronese@gmail.com>